Enhance JIT Tests for IREE Target Validation and Control Flow
- Updated `test_op_norms.cpp`, `test_op_reductions.cpp`, `test_op_rms_norm_expand.cpp`, `test_op_vision.cpp`, and `test_op_where.cpp` to validate JIT operations across all available IREE targets, ensuring numerical consistency with eager execution.
- Added checks for JIT usage and cache statistics to confirm that operations are executed through the IREE backend.
- Enhanced `test_optype_roundtrip.cpp` to ensure proper string representation for additional ops in the interceptor/executor.
- Introduced tests in `test_compile.cpp` to verify correct handling of in-place mutations and execution across varying batch sizes.
- Improved control flow tests in `test_control_flow.cpp` and `test_control_flow_hook_integration.cpp` to ensure accurate execution of conditions and loops across all backends.
- Added symbolic shape inference tests in `test_dynamic_shapes.cpp` to validate behavior for incorrect permutations and reshapes.
- Enhanced extended code generation tests in `test_extended_codegen.cpp` to ensure correct kernel generation for RMS normalization.
- Updated autotune cache validation in `test_jit_lowlevel.cpp` to reject out-of-range algorithm IDs.
- Improved JIT script tests in `test_jit_script.cpp` to ensure multi-argument scripts are callable through vector forward overloads.

+++ include/tenzor/jit/extended_codegen.hpp
@@ -60,7 +60,10 @@ struct ExtendedFusionGroup {
 
     // --- LayerNorm / RMSNorm fields ---
     int norm_axis{-1};                   ///< Normalization axis
-    float eps{1e-5f};
+    // double (not float) so a Float64 LayerNorm/RMSNorm receives a full-precision
+    // eps in the `var + eps` denominator, matching the eager F64 kernel
+    // (JIT-F014). The F32 kernel path narrows it back to float at marshal time.
+    double eps{1e-5};
     bool has_affine{true};               ///< Has gamma/beta parameters
 
     // --- Small MLP fields ---
